Refactor main.rs to start from lib.rs
